FileServer/requirements.txt
Flask
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
psycopg2-binary
cryptography
python-dotenv